requirements.txt
allure-behave==2.5.4
allure-pytest==2.5.4
Appium-Python-Client==0.34
behave==1.2.6
pytest==4.0.2
requests==2.21.0
selenium==3.141.0